feat(update): wait for a rand period in background
Command line and installing new agent are not affected, only
updating the agent in daemon mode always waits.  However, if
the environment variable CLOUDBACKUP_UPDATER_DEBUG is set, the
updater waits for X minutes instead of hours, where X obeys
Gamma(3, 2).  The expectation of X is 6, ~14% of the user can
get X of 4, and the most unlucky user may get X around 20.